requirements.txt
edge_tts==6.1.7
fairseq==0.12.2
faiss_cpu==1.7.4
# gradio==3.47.1
librosa==0.9.1
numpy==1.23.5
praat-parselmouth==0.4.3
pyworld==0.3.4
torchcrepe==0.0.20
